Fixing delete tender by chickig the file existence

+++ out/src/controllers/tender/deleteTender.js
@@ -16,11 +16,13 @@ const deleteTender = async (req, res) => {
         });
         if (!existingTender)
             return res.status(404).json(new response_1.default(false, "Tender not found"));
-        const fileDeletePromises = existingTender.files.map(async (file) => {
-            const filePath = path_1.default.join(__dirname, "../../../public/files/tender", file.file);
-            await promises_1.default.unlink(filePath);
-        });
-        await Promise.all(fileDeletePromises);
+        if (existingTender.files) {
+            const fileDeletePromises = existingTender.files.map(async (file) => {
+                const filePath = path_1.default.join(__dirname, "../../../public/files/tender", file.file);
+                await promises_1.default.unlink(filePath);
+            });
+            await Promise.all(fileDeletePromises);
+        }
         await prismaClient_1.default.tender.delete({
             where: { id: Number(tender_id) },
         });
